+++ bibfor/include/fake/asterc/r4miem.h
@@ -1,2 +1,2 @@
 ! transitional empty interface
-real(kind=4) :: r4miem
+real(kind=8) :: r4miem
