+++ bibfor/include/interfaces/asterc/r4maem.h
@@ -17,6 +17,6 @@
 !
 interface
     function r4maem()
-        real(kind=4) :: r4maem
+        real(kind=8) :: r4maem
     end function r4maem
 end interface
