+++ bibfor/include/interfaces/asterc/r4miem.h
@@ -17,6 +17,6 @@
 !
 interface
     function r4miem()
-        real(kind=4) :: r4miem
+        real(kind=8) :: r4miem
     end function r4miem
 end interface
